third_party/WebKit/LayoutTests/paint/invalidation/text-selection-focus-expected.txt

If this test passes, a frame was scheduled when focusing the window.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scheduled is true
PASS successfullyParsed is true

TEST COMPLETE
Selected area
